python/moose/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.2
django-ckeditor==6.7.1
django-js-asset==2.2.0
idna==3.6
pillow==10.2.0
requests==2.31.0
sqlparse==0.4.4
urllib3==2.2.1
